+++ firebase/functions/main.ts
@@ -1,0 +1,6 @@
+import { config } from 'firebase-functions/v1'
+import { initializeApp } from 'firebase-admin/app'
+
+initializeApp(config().firebase)
+
+export * from 'firebase-functions/v1'
